WIP: Auto adjust BNB balance while using BNB for fees
This patch is used by myself and still being used.

There is still some enhancements like use config to manage it and so forth. I pushed this PR to see if there is anybody interested in it as well. If so, please let me know your comments and suggestions, I'll continue my work to make it complete.